Revision of documents on changes applied during code refactoring.
